+++ modal/__init__.py
@@ -9,15 +9,14 @@ if sys.version_info[:2] >= (3, 13):
 from modal_version import __version__
 
 try:
-    from ._container_io_manager import interact, is_local
     from ._tunnel import Tunnel, forward
-    from ._utils.function_utils import current_function_call_id, current_input_id
     from .app import App, Stub
     from .client import Client
     from .cloud_bucket_mount import CloudBucketMount
     from .cls import Cls
     from .dict import Dict
     from .exception import Error
+    from .execution_context import current_function_call_id, current_input_id, interact, is_local
     from .functions import Function
     from .image import Image
     from .mount import Mount
@@ -73,9 +72,9 @@ __all__ = [
     "exit",
     "forward",
     "is_local",
+    "interact",
     "method",
     "web_endpoint",
     "web_server",
     "wsgi_app",
-    "interact",
 ]
