+++ modal/_utils/function_utils.py
@@ -7,7 +7,6 @@ import sys
 import sysconfig
 import typing
 from collections import deque
-from contextvars import ContextVar
 from enum import Enum
 from pathlib import Path, PurePosixPath
 from typing import Any, AsyncIterator, Callable, List, Literal, Optional, Set, Type
@@ -477,56 +476,3 @@ async def _create_input(args, kwargs, client, idx: Optional[int] = None) -> api_
             input=api_pb2.FunctionInput(args=args_serialized, data_format=api_pb2.DATA_FORMAT_PICKLE),
             idx=idx,
         )
-
-
-_current_input_id: ContextVar = ContextVar("_current_input_id")
-_current_function_call_id: ContextVar = ContextVar("_current_function_call_id")
-
-
-def current_input_id() -> Optional[str]:
-    """Returns the input ID for the current input.
-
-    Can only be called from Modal function (i.e. in a container context).
-
-    ```python
-    from modal import current_input_id
-
-    @stub.function()
-    def process_stuff():
-        print(f"Starting to process {current_input_id()}")
-    ```
-    """
-    try:
-        return _current_input_id.get()
-    except LookupError:
-        return None
-
-
-def current_function_call_id() -> Optional[str]:
-    """Returns the function call ID for the current input.
-
-    Can only be called from Modal function (i.e. in a container context).
-
-    ```python
-    from modal import current_function_call_id
-
-    @stub.function()
-    def process_stuff():
-        print(f"Starting to process input from {current_function_call_id()}")
-    ```
-    """
-    try:
-        return _current_function_call_id.get()
-    except LookupError:
-        return None
-
-
-def _set_current_context_ids(input_id: str, function_call_id: str) -> Callable[[], None]:
-    input_token = _current_input_id.set(input_id)
-    function_call_token = _current_function_call_id.set(function_call_id)
-
-    def _reset_current_context_ids():
-        _current_input_id.reset(input_token)
-        _current_function_call_id.reset(function_call_token)
-
-    return _reset_current_context_ids
