+++ modal/execution_context.py
@@ -1,0 +1,79 @@
+from contextvars import ContextVar
+from typing import Callable, Optional
+
+from modal._container_io_manager import _ContainerIOManager
+from modal._utils.async_utils import synchronize_api
+from modal.exception import InvalidError
+
+
+def is_local() -> bool:
+    """Returns if we are currently on the machine launching/deploying a Modal app
+
+    Returns `True` when executed locally on the user's machine.
+    Returns `False` when executed from a Modal container in the cloud.
+    """
+    return not _ContainerIOManager._singleton
+
+
+async def _interact() -> None:
+    container_io_manager = _ContainerIOManager._singleton
+    if not container_io_manager:
+        raise InvalidError("Interactivity only works inside a Modal container.")
+    else:
+        await container_io_manager.interact()
+
+
+interact = synchronize_api(_interact)
+
+
+def current_input_id() -> Optional[str]:
+    """Returns the input ID for the current input.
+
+    Can only be called from Modal function (i.e. in a container context).
+
+    ```python
+    from modal import current_input_id
+
+    @stub.function()
+    def process_stuff():
+        print(f"Starting to process {current_input_id()}")
+    ```
+    """
+    try:
+        return _current_input_id.get()
+    except LookupError:
+        return None
+
+
+def current_function_call_id() -> Optional[str]:
+    """Returns the function call ID for the current input.
+
+    Can only be called from Modal function (i.e. in a container context).
+
+    ```python
+    from modal import current_function_call_id
+
+    @stub.function()
+    def process_stuff():
+        print(f"Starting to process input from {current_function_call_id()}")
+    ```
+    """
+    try:
+        return _current_function_call_id.get()
+    except LookupError:
+        return None
+
+
+def _set_current_context_ids(input_id: str, function_call_id: str) -> Callable[[], None]:
+    input_token = _current_input_id.set(input_id)
+    function_call_token = _current_function_call_id.set(function_call_id)
+
+    def _reset_current_context_ids():
+        _current_input_id.reset(input_token)
+        _current_function_call_id.reset(function_call_token)
+
+    return _reset_current_context_ids
+
+
+_current_input_id: ContextVar = ContextVar("_current_input_id")
+_current_function_call_id: ContextVar = ContextVar("_current_function_call_id")
