+++ modal/secret.py
@@ -6,12 +6,12 @@ from grpclib import GRPCError, Status
 
 from modal_proto import api_pb2
 
-from ._container_io_manager import is_local
 from ._resolver import Resolver
 from ._utils.async_utils import synchronize_api
 from ._utils.grpc_utils import retry_transient_errors
 from .client import _Client
 from .exception import InvalidError, NotFoundError
+from .execution_context import is_local
 from .object import _get_environment_name, _Object
 
 ENV_DICT_WRONG_TYPE_ERR = "the env_dict argument to Secret has to be a dict[str, Union[str, None]]"
